# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs for Shop Scrape App",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/retailers": {
            "get": {
                "tags": [
                    "Retailers"
                ],
                "summary": "Retrieve all retailers",
                "description": "Fetches a list of all retailers with their associated currencies.",
                "operationId": "ef677653347acacab3fe5655eb27c198",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RetailerResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Retailers"
                ],
                "summary": "Create a retailer",
                "description": "Stores a new retailer in the database.",
                "operationId": "0c02c85dfc6a75fce58175bfd84720ed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Title of the retailer",
                                        "type": "string",
                                        "example": "Retailer 1"
                                    },
                                    "url": {
                                        "description": "URL of the retailer",
                                        "type": "string",
                                        "example": "http://retailer1.com"
                                    },
                                    "currency_id": {
                                        "description": "Currency ID associated with the retailer",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Retailer created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RetailerResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/retailers/{retailer}/products": {
            "get": {
                "tags": [
                    "Retailers"
                ],
                "summary": "Retrieve retailer's products",
                "description": "Fetches the list of products associated with a specific retailer.",
                "operationId": "22a259427e19d37514be1e090a6a036b",
                "parameters": [
                    {
                        "name": "retailer",
                        "in": "path",
                        "description": "ID of the retailer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Retailers"
                ],
                "summary": "Add products to a retailer",
                "description": "Assigns a list of products to a specific retailer.",
                "operationId": "8a98573b3bb1f205160a1e5c1945c249",
                "parameters": [
                    {
                        "name": "retailer",
                        "in": "path",
                        "description": "ID of the retailer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "products": {
                                        "description": "List of product IDs to be assigned",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Products assigned to retailer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RetailerResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/retailers/{retailer}": {
            "put": {
                "tags": [
                    "Retailers"
                ],
                "summary": "Update a retailer",
                "description": "Updates an existing retailer by its ID.",
                "operationId": "84f52f899b9ed757b1f1cee365845c92",
                "parameters": [
                    {
                        "name": "retailer",
                        "in": "path",
                        "description": "ID of the retailer to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Updated title of the retailer",
                                        "type": "string",
                                        "example": "Updated Retailer"
                                    },
                                    "url": {
                                        "description": "Updated URL",
                                        "type": "string",
                                        "example": "http://new-retailer.com"
                                    },
                                    "currency_id": {
                                        "description": "Updated currency ID",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retailer updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RetailerResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Retailers"
                ],
                "summary": "Delete a retailer",
                "description": "Deletes a retailer by its ID.",
                "operationId": "d151bc2b89e222572d28b36b07f4cf07",
                "parameters": [
                    {
                        "name": "retailer",
                        "in": "path",
                        "description": "ID of the retailer to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retailer deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Operation success status",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "Success message",
                                            "type": "string",
                                            "example": "Retailer deleted successfully."
                                        },
                                        "data": {
                                            "description": "Additional response data (empty for delete operation)",
                                            "type": "object",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Retailer not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CurrencyResource": {
                "description": "Currency resource schema",
                "properties": {
                    "id": {
                        "description": "ID of the currency",
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "description": "Currency code",
                        "type": "string",
                        "example": "USD"
                    },
                    "name": {
                        "description": "Currency name",
                        "type": "string",
                        "example": "US Dollar"
                    },
                    "symbol": {
                        "description": "Currency symbol",
                        "type": "string",
                        "example": "$"
                    }
                },
                "type": "object"
            },
...
            "PackSizeResource": {
                "description": "Pack Size resource schema",
                "properties": {
                    "id": {
                        "description": "ID of the pack size",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the pack size",
                        "type": "string",
                        "example": "Small"
                    },
                    "weight": {
                        "description": "Weight of the pack size with unit",
                        "type": "string",
                        "example": "500 g"
                    },
                    "amount": {
                        "description": "Amount of items in the pack",
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
            "ProductResource": {
                "description": "Product resource schema",
                "properties": {
                    "id": {
                        "description": "ID of the product",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Title of the product",
                        "type": "string",
                        "example": "Laptop"
                    },
                    "description": {
                        "description": "Description of the product",
                        "type": "string",
                        "example": "High-performance laptop"
                    },
                    "manufacturer_part_number": {
                        "description": "Manufacturer part number",
                        "type": "string",
                        "example": "LAP12345"
                    },
                    "pack_size": {
                        "$ref": "#/components/schemas/PackSizeResource"
                    },
                    "images": {
                        "description": "List of product images",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductImageResource"
                        }
                    }
                },
                "type": "object"
            },
            "ProductImageResource": {
                "description": "Product image resource schema",
                "properties": {
                    "id": {
                        "description": "ID of the product image",
                        "type": "integer",
                        "example": 1
                    },
                    "file_url": {
                        "description": "URL of the image file",
                        "type": "string",
                        "example": "http://example.com/image.jpg"
                    },
                    "file_name": {
                        "description": "Name of the image file",
                        "type": "string",
                        "example": "image.jpg"
                    }
                },
                "type": "object"
            },
...
            "RetailerResource": {
                "description": "Retailer resource schema",
                "properties": {
                    "id": {
                        "description": "ID of the retailer",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Title or name of the retailer",
                        "type": "string",
                        "example": "Best Buy"
                    },
                    "url": {
                        "description": "URL of the retailer",
                        "type": "string",
                        "example": "http://www.bestbuy.com"
                    },
                    "logo": {
                        "description": "Logo URL of the retailer",
                        "type": "string",
                        "example": "http://www.bestbuy.com/logo.png"
                    },
                    "currency": {
                        "$ref": "#/components/schemas/CurrencyResource"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}